chore(products.js): moved products queries to db/products
